Add files via upload
In this project I have a dataset which includes the votes for US house of representatives and we have to identify whether the voter is a Republican or Democratic and I have implemented two machine learning algorithms KNN and Artificial Neural Networks. 
I have divided the data set into training and test data in the ratio of 80% - 20%. S, I trained the data on KNN and then used the test data for prediction. As, the data is categorical I have converted the data into numerical by using mutate function for KNN and model.matrix function for ANN model. 
I have replaced the NA values with y and n on the basis of count and have used for loop for that. To improve the performance I have implemented 10- fold Cross Validation and performed Automated Parameter Tuning for both KNN and ANN models, where we are finding the accuracy and kappa value for the KNN model and R-squared value for the ANN model to determine the optimal model.